% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hathi-package.R
\docType{package}
\name{hathi-package}
\alias{hathi-package}
\alias{hathi}
\title{Interface to the HathiTrust API.}
\description{
hathi: Interface to the HathiTrust API.
}
\author{
Scott Chamberlain \email{myrmecocystus@gmail.com}
}
