% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hathi_data.r
\name{hathi_data}
\alias{hathi_data}
\title{Hathitrust data collection.}
\usage{
hathi_data()
}
\description{
Hathitrust data collection.
}
\details{
Details:
}
\section{Authentication}{

Get a key at \url{http://babel.hathitrust.org/cgi/kgs/request}
}

\section{More data}{

You can go to their web interface to generate datasets for later download. But wouldn't you
rather do things programatically in R?
}

\examples{
\dontrun{
hathi_data()
}
}
\references{
\url{http://www.hathitrust.org/documents/hathitrust-data-api-v2.pdf}
}
